[BUG] freecells were stack indefinitely
While playing a test game, I accidentally stacked a G-dragon on top of an 8 in the freecells, and I was baffled at where it disappeared to. Moving the 8 out of the freecells into an empty row, it moved “the whole stack” of the 8 and G-dragon into the empty row.

Investigating, I realized this conditional could never be true, because the two conditions are logically exclusive of each other.